+++ Marlin/src/lcd/menu/menu_motion.cpp
@@ -127,16 +127,16 @@ static void _lcd_move_xyz(PGM_P name, AxisEnum axis) {
     #endif
 
     // Get the new position
-    const float diff = float(int16_t(ui.encoderPosition)) * move_menu_scale;
+    const float diff = float(int32_t(ui.encoderPosition)) * move_menu_scale;
     #if IS_KINEMATIC
       manual_move_offset += diff;
-      if (int16_t(ui.encoderPosition) < 0)
+      if (int32_t(ui.encoderPosition) < 0)
         NOLESS(manual_move_offset, min - current_position[axis]);
       else
         NOMORE(manual_move_offset, max - current_position[axis]);
     #else
       current_position[axis] += diff;
-      if (int16_t(ui.encoderPosition) < 0)
+      if (int32_t(ui.encoderPosition) < 0)
         NOLESS(current_position[axis], min);
       else
         NOMORE(current_position[axis], max);
@@ -169,7 +169,7 @@ void lcd_move_z() { _lcd_move_xyz(GET_TEXT(MSG_MOVE_Z), Z_AXIS); }
     if (ui.use_click()) return ui.goto_previous_screen_no_defer();
     if (ui.encoderPosition) {
       if (!ui.processing_manual_move) {
-        const float diff = float(int16_t(ui.encoderPosition)) * move_menu_scale;
+        const float diff = float(int32_t(ui.encoderPosition)) * move_menu_scale;
         #if IS_KINEMATIC
           manual_move_offset += diff;
         #else
